Renames controller files for consistency
Renames the `viewPersonHome.ts` and `viewSentenceBreakdown.ts` files to `viewPersonHomeController.ts` and `viewSentenceBreakdownController.ts` respectively for naming consistency across the project.

Updates imports accordingly.
